ext.relatedArticles.readMore.bootstrap: Convert to 'packageFiles'
This enables use of module.exports and require() within the module,
and makes it easy to bundle additional data and scripts generated
by the server (without needing to subclass FileModule in PHP).

In particular, to bundle mw.config values with the module.
Currently the extension is adding these to the <head> of all
articles which delays rendering of the article content, and the
time to when JS can start to be fetched.

See <https://phabricator.wikimedia.org/T219342#5116773>.

Bug: T219342
Change-Id: I3f174418af8bfd83b0faabe484226e323714cb5b