selenium: Enable one skipped test, delete one failing test
Both tests are disabled for more than a year, see
f318d3acd299a0ccc9c13ee40dd46bc6d7fa14a4.

'ReadMore is not present on Vector' is failing so it is deleted.
'ReadMore is present in Minerva @daily' is passing so it is enabled.

To be sure that the passing test is not flaky, the CI has ran it 100
times, it never failed.

Bug: T280652
Change-Id: I739414c4f3f650ce14e31b6bb2bfdd21efb8ff44