Add a wprov parameter for measuring impact
We would like to know how many users are using this related articles
endpoint. To accomplish define a new wprov parameter (on wikitech)
and use it here. wprov is specifically defined to avoid caching
issues, this should be safe to attach to any page view to track
that page views provenance.

Bug: T358351
Change-Id: I58b450eb632a721a39b627cc34f30eb26e67be59