SECURITY: Escape card details
The title and description are probably the most important, but the rest
seems prudent to escape too.

mw.html is part of the mediawiki.base module, so no new ResourceLoader
dependency should be necessary. However, it’s not currently part of
@wikimedia/types-wikimedia, so we need to ts-ignore the calls to it for
now; to simplify this, extract an escapeHtml() function.

To make the tests work, add a URL to the cards there.

Bug: T396413
Change-Id: Ib3707e8cc4106d914260c9f3eadf651756705afe